+++ src/test/java/org/openlmis/referencedata/testbuilder/ProcessingScheduleDataBuilder.java
@@ -16,6 +16,7 @@
 package org.openlmis.referencedata.testbuilder;
 
 import org.openlmis.referencedata.domain.ProcessingSchedule;
+
 import java.time.ZonedDateTime;
 import java.util.UUID;