+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Breaking changes:
 parameters to get users endpoint
 * [OLMIS-3448](https://openlmis.atlassian.net/browse/OLMIS-3448): Paginate minimal facilities endpoint
 * [OLMIS-2695](https://openlmis.atlassian.net/browse/OLMIS-2695): Paginate get all processing schedules endpoint.
+* [OLMIS-2695](https://openlmis.atlassian.net/browse/OLMIS-2695): Merged processing period search endpoints with get all periods endpoint. Convert endpoint to be pageable.
 
 New functionality added in a backwards-compatible manner:
 * [OLMIS-3782](https://openlmis.atlassian.net/browse/OLMIS-3782): Ability to set up program to skip requisition authorization step